% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASPIRE.obj}
\alias{ASPIRE.obj}
\title{Objective table for an ASPIRE analysis}
\format{A data frame with columns:
\describe{
 \item{Variables}{variables names}
 \item{Utility}{utility functions to be used for each variables}
 \item{Obj1}{the weights of each variable for objective Obj1}
}}
\source{
Fake data, just to illustrate the ASPIRE package
}
\usage{
ASPIRE.obj
}
\description{
Objective table for an ASPIRE analysis
}
\examples{
ASPIRE.obj

}
\keyword{datasets}
