% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASPIRE.proj}
\alias{ASPIRE.proj}
\title{Project table for an ASPIRE analysis}
\format{A data frame with columns:
\describe{
 \item{Objectives}{objectives names}
 \item{Stak1}{the weights of each objectives for the stakeholder Stak1}
}}
\source{
Fake data, just to illustrate the ASPIRE package
}
\usage{
ASPIRE.proj
}
\description{
Project table for an ASPIRE analysis
}
\examples{
ASPIRE.proj

}
\keyword{datasets}
