% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASPIRE.var}
\alias{ASPIRE.var}
\title{Variable table for an ASPIRE analysis}
\format{A data frame with columns:
\describe{
 \item{Var1}{variable values}
 \item{Var1_Ref}{reference values for the Var1 variable}
}}
\source{
Fake data, just to illustrate the ASPIRE package
}
\usage{
ASPIRE.var
}
\description{
Variable table for an ASPIRE analysis
}
\examples{
ASPIRE.var

}
\keyword{datasets}
