% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Obj.barplot.all}
\alias{Obj.barplot.all}
\title{objective barplot}
\usage{
Obj.barplot.all(
  objective_name,
  project_all,
  plot.ref = TRUE,
  ylim = NULL,
  las.x = 1,
  cex.x = 1,
  main = NULL,
  ...
)
}
\arguments{
\item{objective_name}{an objective name}

\item{project_all}{a project object from \code{\link{ASPIRE_all}}}

\item{plot.ref}{a logical value indicating whether the reference should be drawn or not}

\item{ylim}{limits for the y axis}

\item{las.x}{axis label horizontal (1) or perpendicular to the axis (2)}

\item{cex.x}{magnification to be used for the axis label}

\item{main}{an overall title for the plot}

\item{...}{other arguments from \code{\link{barplot}}}
}
\description{
after \code{\link{ASPIRE_all}}
}
