% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Obj.barplot}
\alias{Obj.barplot}
\title{Objective barplot}
\usage{
Obj.barplot(
  obj,
  plot.ref = TRUE,
  ylim = NULL,
  las.x = 1,
  cex.x = 1,
  main = NULL,
  ...
)
}
\arguments{
\item{obj}{an Objective object created with \code{\link{Objective}}}

\item{plot.ref}{a logical value indicating whether the reference should be drawn or not}

\item{ylim}{limits for the y axis}

\item{las.x}{axis label horizontal (1) or perpendicular to the axis (2)}

\item{cex.x}{magnification to be used for the axis label}

\item{main}{an overall title for the plot}

\item{...}{other arguments from \code{\link{barplot}}}
}
\description{

}
\examples{
# ASPIRE.var and ASPIRE.obj are data.frame included in the ASPIRE package
# creation of the different Variable objects:
for(i in 1:nrow(ASPIRE.obj))
{
  eval(parse(text = paste(names(ASPIRE.var)[i],
                          "<-Variable(values = ASPIRE.var[,i],ref = ASPIRE.var[,i+(ncol(ASPIRE.var)/2)],utility = ASPIRE.obj$Utility[i])")))
}
# creation of an Objective object:
Obj1 <- Objective(var_names = c("Div_Butterflies","Div_Plant","Compo_Butterflies","Compo_Plant"),
                  var_weight = c(2,1,1,2))
# plot:
Obj.barplot(obj = Obj1)
Obj.barplot(obj = Obj1,plot.ref = FALSE, ylim = c(0,2), col = "cadetblue",las.x = 2, cex.x = 0.7)
}
\seealso{
\code{\link{Obj.barplot.all}}
}
