% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Obj.radar.plot.all}
\alias{Obj.radar.plot.all}
\title{Objective radar diagram}
\usage{
Obj.radar.plot.all(
  objective_name,
  project_all,
  limited = TRUE,
  transp = 0.5,
  col = "c7",
  main = "Variables scores",
  plot.ref = TRUE,
  ...
)
}
\arguments{
\item{objective_name}{an objective name}

\item{project_all}{a project object from \code{\link{ASPIRE_all}}}

\item{limited}{a logical value indicating whether transformed values above 1 should be plotted (therefore avoiding plot distortion)}

\item{transp}{a value between 0 and 1 indicating transparancy of the radar polygon}

\item{col}{radar polygon color, can be any color managed by R, but \code{transp} won't work if other values "c1", "c..", "c8" are used, corresponding to basic R colors from 1 to 8}

\item{main}{text for main title}

\item{plot.ref}{a logical value indicating whether the reference should be drawn or not}

\item{...}{any other arguments applicable to \code{\link{fmsb::radarchart}}}
}
\description{
draw a radar diagram with variables scores of one objective after \code{\link{ASPIRE_all}}
}
\details{
The function uses the function \code{\link{radarchart}} from the package \code{\link{fmsb}} Minato Nakazawa (2014). fmsb: Functions for medical statistics book with some demographic data. R package version 0.7.0.
}
\examples{
# ASPIRE.var, ASPIRE.obj and ASPIRE.proj are data.frame included in the ASPIRE package
# creation of an ASPIRE_all object:
Proj_all <- ASPIRE_all(variable_df = ASPIRE.var,objective_df = ASPIRE.obj,project_df = ASPIRE.proj)
Obj.radar.plot.all(objective_name = "Functioning",project_all = Proj_all)
Obj.radar.plot.all(objective_name = "Functioning",project_all = Proj_all,
                   limited = FALSE,transp = 0.1,col = "c6",main = "Birds",plot.ref = FALSE)
}
