% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Obj.radar.plot}
\alias{Obj.radar.plot}
\title{Objective radar diagram}
\usage{
Obj.radar.plot(
  obj,
  limited = TRUE,
  transp = 0.5,
  col = "c7",
  main = "Variables scores",
  plot.ref = TRUE,
  ...
)
}
\arguments{
\item{obj}{an objective object created with \code{\link{Objective}}}

\item{limited}{a logical value indicating whether transformed values above 1 should be plotted (therefore avoiding plot distortion)}

\item{transp}{a value between 0 and 1 indicating transparancy of the radar polygon}

\item{col}{radar polygon color, can be any color managed by R, but \code{transp} won't work if other values "c1", "c..", "c8" are used, corresponding to basic R colors from 1 to 8}

\item{main}{text for main title}

\item{plot.ref}{a logical value indicating whether the reference should be drawn or not}

\item{...}{any other arguments applicable to \code{\link{fmsb::radarchart}}}
}
\description{
draw a radar diagram with variables scores of one objective
}
\details{
The function uses the function \code{\link{radarchart}} from the package \code{\link{fmsb}} Minato Nakazawa (2014). fmsb: Functions for medical statistics book with some demographic data. R package version 0.7.0.
}
\examples{
# ASPIRE.var and ASPIRE.obj are data.frame included in the ASPIRE package
# creation of the different Variable objects:
for(i in 1:nrow(ASPIRE.obj))
{
  eval(parse(text = paste(names(ASPIRE.var)[i],
                          "<-Variable(values = ASPIRE.var[,i],ref = ASPIRE.var[,i+(ncol(ASPIRE.var)/2)],utility = ASPIRE.obj$Utility[i])")))
}
# creation of an Objective object:
Obj1 <- Objective(var_names = c("Div_Butterflies","Div_Plant","Compo_Butterflies","Compo_Plant"),
                  var_weight = c(2,1,1,2))
# plot:
Obj.radar.plot(obj = Obj1)
Obj.radar.plot(obj = Obj1,limited = FALSE, transp = 0.1,
               col = "c3",main = "Nice title",plot.ref = FALSE,vlcex=0.7)
}
\seealso{
\code{\link{Obj.radar.plot.all}}
}
