% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Proj.all.radar.plot}
\alias{Proj.all.radar.plot}
\title{All scores radar diagram}
\usage{
Proj.all.radar.plot(
  proj,
  transp = 0.5,
  col = NULL,
  AllInOne = TRUE,
  limited = TRUE,
  main = "",
  plot.ref = TRUE,
  cex.lab = 1
)
}
\arguments{
\item{proj}{project name defined with \code{\link{Project}} or \code{\link{ASPIRE_all}}}

\item{transp}{a value between 0 and 1 indicating transparancy of the radar polygon}

\item{col}{polygon color, can be any color managed by R, but \code{transp} won't work if other values than "c1", "c..", "c8" are used, corresponding to basic R colors from 1 to 8. If only one color is provided and \code{AllInOne=FALSE}, color is repeated  for all objectives}

\item{AllInOne}{a logical value indicating whether all the variables scores should be displeyd on one radardiagram or assembled by objectives}

\item{limited}{a logical value indicating whether transformed values above 1 should be plotted (therefore avoiding plot distortion)}

\item{main}{text for main title, only applicable if \code{AllInOne=TRUE}}

\item{plot.ref}{a logical value indicating whether the reference should be drawn or not}

\item{cex.lab}{magnification to be used for the labels}
}
\description{
displays either one or several radar diagrams of all variables scores
}
\details{
The function uses the function \code{\link{radarchart}} from the package \code{\link{fmsb}} Minato Nakazawa (2014). fmsb: Functions for medical statistics book with some demographic data. R package version 0.7.0.
}
