% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Proj.obj.barplot}
\alias{Proj.obj.barplot}
\title{Proj.obj.barplot}
\usage{
Proj.obj.barplot(proj, ylim = NULL, las.x = 1, cex.x = 1, ...)
}
\arguments{
\item{proj}{project name defined with \code{\link{Project}} or \code{\link{ASPIRE_all}}}

\item{ylim}{limits for the y axis}

\item{las.x}{axis label horizontal (1) or perpendicular to the axis (2)}

\item{cex.x}{magnification to be used for the axis label}

\item{...}{other arguments applicable to \code{\link{barplot}}}
}
\description{
displays a barplot of objectives scores
}
