% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Proj.radar.plot}
\alias{Proj.radar.plot}
\title{Objective scores radar diagram}
\usage{
Proj.radar.plot(
  proj,
  limited = TRUE,
  transp = 0.5,
  col = "c7",
  main = "Objectives scores",
  ...
)
}
\arguments{
\item{proj}{project name defined with \code{\link{Project}} or \code{\link{ASPIRE_all}}}

\item{limited}{a logical value indicating whether transformed values above 1 should be plotted (therefore avoiding plot distortion)}

\item{transp}{a value between 0 and 1 indicating transparancy of the radar polygon}

\item{col}{radar polygon color, can be any color managed by R, but \code{transp} won't work if other values "c1", "c..", "c8" are used, corresponding to basic R colors from 1 to 8}

\item{main}{text for main title}

\item{...}{any other arguments applicable to \code{\link{fmsb::radarchart}}}
}
\description{
displays radar diagram of objectives scores
}
\details{
The function uses the function \code{\link{radarchart}} from the package \code{\link{fmsb}} Minato Nakazawa (2014). fmsb: Functions for medical statistics book with some demographic data. R package version 0.7.0.
}
