% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Project}
\alias{Project}
\title{Creation of the ASPIRE Project object}
\usage{
Project(obj_name, stakeholders, obj_weight)
}
\arguments{
\item{obj_name}{a vector \code{c()} including objectives names defined with \code{\link{Objective}}}

\item{stakeholders}{a dataframe \code{data.frame()} of objectives weights with a column for each staekholder (in the same order as  \code{stakeholders}) and the objectives as row  (in the same order than \code{obj_name})}

\item{obj_weight}{a vector \code{c()} of weights for all the objectives}
}
\value{
\item{name_objective}{objectives names given in the obj_name argument}

\item{DF_Obj_Proj}{a dataframe with a row by objective and including the values calculated with \code{\link{Objective}}}

\item{Proj_WMean}{weighted mean of objectives scores}

\item{Proj_Wsem}{weighted mean of objectives scores errors}

\item{Proj_inf}{weighted mean of objectives scores minus error (useful for graphical purpose)}

\item{Proj_sup}{weighted mean of objectives scores plus error (useful for graphical purpose)}

\item{summary}{A summary dataframe of Project object}

\item{Variables}{A list of objects obtained from \code{\link{Variable}}, one per variable}

\item{Objectives}{A list of objects obtained from \code{\link{Objective}}, one per objective}
}
\description{
Creates an ASPIRE Project object
}
\seealso{
\code{\link{Transf_ASPIRE}}, \code{\link{Variable}}, \code{\link{Objective}}, \code{\link{Project}}
}
