% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Transf_ASPIRE}
\alias{Transf_ASPIRE}
\title{Transformation of variable values into variable scores}
\usage{
Transf_ASPIRE(
  VAL,
  UTIL,
  VAL_MEAN_REF,
  ALPHA = NULL,
  XVAL = NULL,
  YSCO = NULL,
  NEG = FALSE
)
}
\arguments{
\item{VAL}{variable value, it should only contains one value}

\item{UTIL}{utility function: "continuous" linear relationship, score can be above 1 ;"threshold" linear relationship,
score does not exceed 1 ;"thresholdalpha" linear relationship until ALPHA, then 1 ; "perso", linear relationships between points
of which x (values) and y (scores) are given in XVAL et YSCO parameters ; "step", 0 if below reference, 1 when above ; "stepalpha", 0 if below ALPHA, 1 when above ;
"stepperso", step relationship between points of which x (values) and y (scores) are given in XVAL et YSCO parameters ; "hump", similar to "continuous" relationship
the reference, then decreasing with the same negative slope}

\item{VAL_MEAN_REF}{mean value for the reference}

\item{ALPHA}{threshold value (cf UTIL)}

\item{XVAL}{x value for "perso" and "stepperso" utility fonctions (cf UTIL)}

\item{YSCO}{y value for "perso" and "stepperso" utility fonctions (cf UTIL)}

\item{NEG}{if FALSE (by default), score value can't be negative, it is transformed into 0}
}
\value{
a score value
}
\description{
Calculates score value after transformation relative to the reference value and according to the chosen utility function.
}
\examples{
VAL_tests <- seq(from = -4, to = 16, by = 1)
VAL_MEAN_REF_tests <- 6
XVAL_tests <- c(1, 6, 10)
YSCO_tests <- c(0, 1, 0.5)

Transf_ASPIRE(VAL = VAL_tests, UTIL = "continuous",VAL_MEAN_REF = VAL_MEAN_REF_tests)
Transf_ASPIRE(VAL = VAL_tests, UTIL = "continuous",VAL_MEAN_REF = VAL_MEAN_REF_tests, NEG = TRUE)
Transf_ASPIRE(VAL = VAL_tests, UTIL = "threshold",VAL_MEAN_REF = VAL_MEAN_REF_tests)
Transf_ASPIRE(VAL = VAL_tests, UTIL = "thresholdalpha",VAL_MEAN_REF = VAL_MEAN_REF_tests,ALPHA = 4)
Transf_ASPIRE(VAL = VAL_tests, UTIL = "hump",VAL_MEAN_REF = VAL_MEAN_REF_tests)
Transf_ASPIRE(VAL = VAL_tests, UTIL = "step",VAL_MEAN_REF = VAL_MEAN_REF_tests)
Transf_ASPIRE(VAL = VAL_tests, UTIL = "perso",VAL_MEAN_REF = VAL_MEAN_REF_tests,XVAL = XVAL_tests, YSCO = YSCO_tests)
}
