% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Var.barplot.all}
\alias{Var.barplot.all}
\title{Variable barplot}
\usage{
Var.barplot.all(
  variable_name,
  project_all,
  plot.ref = TRUE,
  ylim = NULL,
  transf = FALSE,
  main = NULL,
  ...
)
}
\arguments{
\item{variable_name}{a variable name}

\item{project_all}{a project object from \code{\link{ASPIRE_all}}}

\item{plot.ref}{a logical value indicating whether the reference should be drawn or not}

\item{ylim}{limits for the y axis}

\item{transf}{a logical value indicating whether the values should be transformated (into scores) or not}

\item{main}{an overall title for the plot}

\item{...}{other arguments applicable to \code{\link{barplot}}}
}
\description{
barplot of variable mean value or score, after \code{\link{ASPIRE_all}}
}
