% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Var.barplot}
\alias{Var.barplot}
\title{Variable barplot}
\usage{
Var.barplot(
  variab,
  plot.ref = TRUE,
  ylim = NULL,
  transf = FALSE,
  main = NULL,
  ...
)
}
\arguments{
\item{variab}{a variable object created with \code{\link{Variable}}}

\item{plot.ref}{a logical value indicating whether the reference should be drawn or not}

\item{ylim}{limits for the y axis}

\item{transf}{a logical value indicating whether the values should be transformated (into scores) or not}

\item{main}{an overall title for the plot}

\item{...}{other arguments applicable to \code{\link{barplot}}}
}
\description{
barplot of variable mean value or score
}
