% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Variable}
\alias{Variable}
\title{Creation of the ASPIRE Variable object}
\usage{
Variable(
  values,
  ref,
  utility = "continuous",
  ALPHA = NULL,
  XVAL = NULL,
  YSCO = NULL,
  NEG = FALSE
)
}
\arguments{
\item{values}{a value or vector of values for the variable, if NA, they are not taken into account}

\item{ref}{a value or vector of values for the reference, if NA, they are not taken into account}

\item{utility}{one category of utility function (cf function \code{\link{Transf_ASPIRE}})}

\item{ALPHA}{threshold value ALPHA (cf function \code{\link{Transf_ASPIRE}})}

\item{XVAL}{x values for the "perso" utility function (cf function \code{\link{Transf_ASPIRE}})}

\item{YSCO}{y scores for the "perso" utility function (cf function \code{\link{Transf_ASPIRE}})}

\item{NEG}{if FALSE (default), score values cannot be negative, it is then 0 (cf function \code{\link{Transf_ASPIRE}})}
}
\value{
\item{calculated_V}{a list of values or calculated values neede for graphical purposes}

\item{summary}{Summary of the Variable object attribute}
}
\description{
Creates an ASPIRE Variable object
}
\examples{
# ASPIRE.var is a data.frame included in the ASPIRE package
Var1 <- Variable(values = ASPIRE.var$Div_Butterflies,
ref = ASPIRE.var$Div_Butterflies_Ref,
utility = "continuous")
}
\seealso{
\code{\link{Transf_ASPIRE}}, \code{\link{Variable}}, \code{\link{Objective}}, \code{\link{Project}}
}
