% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Wheelscores.all}
\alias{Wheelscores.all}
\title{Score wheel}
\usage{
Wheelscores.all(
  project_all,
  col.score = "#FDC919",
  col.null = "white",
  col.lines = "grey90",
  col.border = "#FFEEC3",
  cex.sup = 1,
  cex.low = 0.7,
  col.low = 1,
  col.sup = 1
)
}
\arguments{
\item{project_all}{a project object from \code{\link{ASPIRE_all}}}

\item{col.score}{The color of achieved scores}

\item{col.null}{The color of unachieved scores}

\item{col.lines}{The color of lines and circles}

\item{col.border}{The color of the border containing categories labels}

\item{cex.sup}{An expansion factor for categories labels}

\item{cex.low}{An expansion factor for scores labels}

\item{col.low}{text color for scores labels}

\item{col.sup}{text color for categories labels}
}
\description{
Draw a score wheel inspired from the recovery wheel of SER Standards (McDonald et al., 2016), after \code{\link{ASPIRE_all}}
}
\details{
uses function from plotrix package Lemon, J. (2006) Plotrix: a package in the red light district of R. R-News, 6(4): 8-12.
}
\examples{
# ASPIRE.var, ASPIRE.obj and ASPIRE.proj are data.frame included in the ASPIRE package
# creation of an ASPIRE_all object:
Proj_all <- ASPIRE_all(variable_df = ASPIRE.var,objective_df = ASPIRE.obj,project_df = ASPIRE.proj)
Wheelscores.all(project_all = Proj_all)
Wheelscores.all(project_all = Proj_all, col.score = "cadetblue",
                col.null = 1,col.lines = "green4",col.border = "firebrick",
                cex.sup = 1.1, cex.low = 0.8, col.low = "white", col.sup = "grey95")
}
