% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Wheelscores}
\alias{Wheelscores}
\title{Score wheel}
\usage{
Wheelscores(
  lowercat,
  supercat,
  scores,
  col.score = "#a1d99b",
  col.null = "white",
  col.lines = "black",
  col.border = "grey",
  cex.sup = 1,
  cex.low = 0.7,
  val.max = 5,
  col.low = 1,
  col.sup = 1
)
}
\arguments{
\item{lowercat}{A vector of scores labels}

\item{supercat}{A vector of categories labels}

\item{scores}{A numeric vector of scores}

\item{col.score}{The color of achieved scores}

\item{col.null}{The color of unachievde scores}

\item{col.lines}{The color of lines and circles}

\item{col.border}{The color of the border containing categories labels}

\item{cex.sup}{An expansion factor for categories labels}

\item{cex.low}{An expansion factor for scores labels}

\item{val.max}{theoretical score maximum value (5 by defaut)}

\item{col.low}{text color for scores labels}

\item{col.sup}{text color for categories labels}
}
\description{
Draw a score wheel inspired from the recovery wheel of SER Standards (McDonald et al., 2016)
}
\details{
uses function from plotrix package Lemon, J. (2006) Plotrix: a package in the red light district of R. R-News, 6(4): 8-12.
}
\examples{
# data to draw the same scores as in McDonald et al., 2016
sousTcat<-c("Spatial mosaic","All trophic levels","All strata present",
"No undesirable species","Desirable animals","Desirable plants",
"Water chemo-physical","Substrate chemical","Substrate physical",
"Contamination","Invasive species","Over-utilization",
"Habitat links","Gene flows","Landscape flows",
"Resilience/recruitment","Habitat & interactions","Productivity/cycling")
superTcat<-c(rep(c("STRUCTURAL DIVERSITY","SPECIES COMPOSITION", "PHYSICAL CONDITIONS", 
"ABSENCE OF THREATS", "EXTERNAL EXCHANGES", "ECOSYSTEM FUNCTION"),each=3))
scoTres<-c(2,2,3,4,2,4,4,4,4,5,3,5,3,2,2,0,2,2)

Wheelscores(lowercat = sousTcat,supercat = superTcat, scores = scoTres, col.lines = "cadetblue4",col.score = "darkolivegreen2",
col.border = "gray97",cex.sup = 1.2,cex.low = 0.6)
}
