% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Obj.barplot.all}
\alias{Obj.barplot.all}
\title{Objective barplot}
\usage{
Obj.barplot.all(
  objective_name,
  project_all,
  plot.ref = TRUE,
  ylim = NULL,
  las.x = 1,
  cex.x = 1,
  main = NULL,
  ...
)
}
\arguments{
\item{objective_name}{an objective name}

\item{project_all}{a project object from \code{\link{ASPIRE_all}}}

\item{plot.ref}{a logical value indicating whether the reference should be drawn or not}

\item{ylim}{limits for the y axis}

\item{las.x}{axis label horizontal (1) or perpendicular to the axis (2)}

\item{cex.x}{magnification to be used for the axis label}

\item{main}{an overall title for the plot}

\item{...}{other arguments from \code{\link{barplot}}}
}
\description{
after \code{\link{ASPIRE_all}}
}
\examples{
# ASPIRE.var, ASPIRE.obj and ASPIRE.proj are data.frame included in the ASPIRE package
# creation of an ASPIRE_all object:
Proj_all <- ASPIRE_all(variable_df = ASPIRE.var,objective_df = ASPIRE.obj,project_df = ASPIRE.proj)
Obj.barplot.all(objective_name = "Functioning",project_all = Proj_all)
Obj.barplot.all(objective_name = "Functioning",project_all = Proj_all,
                plot.ref = FALSE,ylim = c(0,2),las.x = 2,cex.x = 0.6,main = "Birds")
}
