% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Objective}
\alias{Objective}
\title{Creation of the ASPIRE Objective object}
\usage{
Objective(var_names, var_weight)
}
\arguments{
\item{var_names}{a vector \code{c()} including variables names defined with \code{\link{Variable}}}

\item{var_weight}{a vector \code{c()} of weights for each of these variables}
}
\value{
\item{summary}{Summary of the Objective object attribute}

\item{name_var}{variables names given in the var_names argument}

\item{DF_Var_Obj}{A dataframe about the objectives useful for graphical purpose}

\item{Obj_WMean}{Mean objective score}

\item{Obj_Wsem}{Error for the objective}

\item{Obj_inf}{Mean objective score minus error (useful for graphical purpose)}

\item{Obj_sup}{Mean objective score plus error (useful for graphical purpose)}

\item{V}{a list of variables objects}
}
\description{
Creates an ASPIRE Objective object
}
\seealso{
\code{\link{Transf_ASPIRE}}, \code{\link{Variable}}, \code{\link{Objective}}, \code{\link{Project}}
}
