% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Proj.complete.barplot}
\alias{Proj.complete.barplot}
\title{Proj.complete.barplot}
\usage{
Proj.complete.barplot(
  proj,
  col.stak = NULL,
  col.obj = NULL,
  col.var = NULL,
  plot.ref = TRUE,
  las.x = 1,
  cex.x = 1
)
}
\arguments{
\item{proj}{project name defined with \code{\link{Project}} or \code{\link{ASPIRE_all}}}

\item{col.stak}{color of stakeholders plot}

\item{col.obj}{color of objectives plot}

\item{col.var}{color of variable scores plot (repeated if only one color is provided)}

\item{plot.ref}{a logical value indicating whether the reference should be drawn or not}

\item{las.x}{axis label horizontal (1) or perpendicular to the axis (2)}

\item{cex.x}{magnification to be used for the axis label}
}
\description{
displays a barplot of stakeholders project scores
}
