% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Proj.obj.barplot}
\alias{Proj.obj.barplot}
\title{Project objectives plot}
\usage{
Proj.obj.barplot(proj, ylim = NULL, las.x = 1, cex.x = 1, ...)
}
\arguments{
\item{proj}{project name defined with \code{\link{Project}} or \code{\link{ASPIRE_all}}}

\item{ylim}{limits for the y axis}

\item{las.x}{axis label horizontal (1) or perpendicular to the axis (2)}

\item{cex.x}{magnification to be used for the axis label}

\item{...}{other arguments applicable to \code{\link{barplot}}}
}
\description{
displays a barplot of objectives scores
}
\examples{
# ASPIRE.var, ASPIRE.obj and ASPIRE.proj are data.frame included in the ASPIRE package
# creation of an ASPIRE_all object:
Proj_all <- ASPIRE_all(variable_df = ASPIRE.var,objective_df = ASPIRE.obj,project_df = ASPIRE.proj)
Proj.obj.barplot(proj = Proj_all)
Proj.obj.barplot(proj = Proj_all, ylim = c(0,2), las.x = 2, cex.x = 0.8)
}
