% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Project}
\alias{Project}
\title{Creation of the ASPIRE Project object}
\usage{
Project(obj_name, stakeholders, obj_weight)
}
\arguments{
\item{obj_name}{a vector \code{c()} including objectives names defined with \code{\link{Objective}}}

\item{stakeholders}{a vector \code{c()} of stakeholders names}

\item{obj_weight}{a dataframe \code{data.frame()} of objectives weights with a column for each staekholder (in the same order as  \code{stakeholders}) and the objectives as row  (in the same order than \code{obj_name})}
}
\value{
\item{name_objective}{objectives names given in the obj_name argument}

\item{DF_Obj_Proj}{a dataframe with a row by objective and including the values calculated with \code{\link{Objective}}}

\item{Proj_WMean}{weighted mean of objectives scores}

\item{Proj_Wsem}{weighted mean of objectives scores errors}

\item{Proj_inf}{weighted mean of objectives scores minus error (useful for graphical purpose)}

\item{Proj_sup}{weighted mean of objectives scores plus error (useful for graphical purpose)}

\item{summary}{A summary dataframe of Project object}

\item{Variables}{A list of objects obtained from \code{\link{Variable}}, one per variable}

\item{Objectives}{A list of objects obtained from \code{\link{Objective}}, one per objective}
}
\description{
Creates an ASPIRE Project object
}
\examples{
# ASPIRE.var, ASPIRE.obj and ASPIRE.proj are data.frame included in the ASPIRE package
# creation of the different Variable objects:
for(i in 1:nrow(ASPIRE.obj))
{
  eval(parse(text = paste(names(ASPIRE.var)[i],
                          "<-Variable(values = ASPIRE.var[,i],ref = ASPIRE.var[,i+(ncol(ASPIRE.var)/2)],utility = ASPIRE.obj$Utility[i])")))
}
# creation of the different Objective objects
for(i in 1:(ncol(ASPIRE.obj)-2))
{
  eval(parse(text = paste(names(ASPIRE.obj)[2+i],"<-Objective(var_names = as.vector(ASPIRE.obj$Variables[ASPIRE.obj[2+i]!=0]),var_weight = ASPIRE.obj[ASPIRE.obj[2+i]!=0,2+i])")))
}
# creation of a Project object:
Proj <- Project(obj_name = c("Biodiversity","Functioning","Production"),
                stakeholders = c("Naturalist_Asso","Env_Managers","Farmers"),
                obj_weight = data.frame(c(6,3,1),c(2,2,1),c(1,1,3)))
}
\seealso{
\code{\link{Transf_ASPIRE}}, \code{\link{Variable}}, \code{\link{Objective}}, \code{\link{Project}}
}
