% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPIRE_functions.R
\name{Var.barplot.all}
\alias{Var.barplot.all}
\title{Variable barplot}
\usage{
Var.barplot.all(
  variable_name,
  project_all,
  plot.ref = TRUE,
  ylim = NULL,
  transf = FALSE,
  main = NULL,
  ...
)
}
\arguments{
\item{variable_name}{a variable name}

\item{project_all}{a project object from \code{\link{ASPIRE_all}}}

\item{plot.ref}{a logical value indicating whether the reference should be drawn or not}

\item{ylim}{limits for the y axis}

\item{transf}{a logical value indicating whether the values should be transformated (into scores) or not}

\item{main}{an overall title for the plot}

\item{...}{other arguments applicable to \code{\link{barplot}}}
}
\description{
barplot of variable mean value or score, after \code{\link{ASPIRE_all}}
}
\examples{
# ASPIRE.var, ASPIRE.obj and ASPIRE.proj are data.frame included in the ASPIRE package
# creation of an ASPIRE_all object:
Proj_all <- ASPIRE_all(variable_df = ASPIRE.var,objective_df = ASPIRE.obj,project_df = ASPIRE.proj)
Var.barplot.all(variable_name = "Soil_Resp", project_all = Proj_all)
Var.barplot.all(variable_name = "Soil_Resp", project_all = Proj_all, 
                main = "Soil respiration", col="cadetblue", plot.ref = FALSE, ylim=c(0,2))
}
